# specs/001-todo-app/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "오늘의 할 일 API",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "메인 화면 조회",
        "responses": {
          "200": {
            "description": "HTML 화면"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "상태 확인",
        "responses": {
          "200": {
            "description": "정상 상태"
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "summary": "할 일 목록 조회",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["all", "active", "completed"]
            },
            "required": false,
            "description": "필터 상태"
          }
        ],
        "responses": {
          "200": {
            "description": "할 일 목록"
          }
        }
      },
      "post": {
        "summary": "할 일 생성",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 할 일"
          },
          "422": {
            "description": "빈 제목"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "patch": {
        "summary": "할 일 부분 수정",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 할 일"
          },
          "404": {
            "description": "존재하지 않는 ID"
          }
        }
      },
      "delete": {
        "summary": "할 일 삭제",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 성공"
          },
          "404": {
            "description": "존재하지 않는 ID"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodoRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "required": ["completed"],
        "properties": {
          "completed": {
            "type": "boolean"
          }
        }
      }
    }
  }
}